refactor: split code (#5)
* feat: split momentum code

* feat: split volatility code

* feat: split trend code

* rf: clippy and fmt

* doc: rewrite readme

* rf: split volume code

* rf: have module naming consistency accross project
