+++ CHANGELOG.md
@@ -1,24 +1,31 @@
 # Changelog
 
-## [Unreleased](https://github.com/Lsh0x/tars/tree/HEAD)
+## [Unreleased](https://github.com/Lsh0x/rsta/tree/HEAD)
 
-[Full Changelog](https://github.com/Lsh0x/tars/compare/f9d6d027ed2e0164435e56e6c12f0184422a79c8...HEAD)
+[Full Changelog](https://github.com/Lsh0x/rsta/compare/v0.0.2...HEAD)
 
-**Merged pull requests:**
+No unreleased changes yet.
 
-- fix: cron for sync template [\#16](https://github.com/Lsh0x/tars/pull/16) ([Lsh0x](https://github.com/Lsh0x))
-- feat: add workflow to sync template [\#15](https://github.com/Lsh0x/tars/pull/15) ([Lsh0x](https://github.com/Lsh0x))
-- .feat: update codecove version [\#14](https://github.com/Lsh0x/tars/pull/14) ([Lsh0x](https://github.com/Lsh0x))
-- feat: rework codecov workflow [\#13](https://github.com/Lsh0x/tars/pull/13) ([Lsh0x](https://github.com/Lsh0x))
-- feat: rework codecov workflow [\#12](https://github.com/Lsh0x/tars/pull/12) ([Lsh0x](https://github.com/Lsh0x))
-- fix: libsso prob du latest ubuntu [\#11](https://github.com/Lsh0x/tars/pull/11) ([Lsh0x](https://github.com/Lsh0x))
-- feat: upgrade version of actions cache [\#8](https://github.com/Lsh0x/tars/pull/8) ([Lsh0x](https://github.com/Lsh0x))
-- fix: change last commit badge from master to main [\#5](https://github.com/Lsh0x/tars/pull/5) ([Lsh0x](https://github.com/Lsh0x))
-- fix: github action coverage workflow [\#4](https://github.com/Lsh0x/tars/pull/4) ([Lsh0x](https://github.com/Lsh0x))
-- Fix/transform template into lib [\#3](https://github.com/Lsh0x/tars/pull/3) ([Lsh0x](https://github.com/Lsh0x))
-- feat: update github workflow for changelog [\#2](https://github.com/Lsh0x/tars/pull/2) ([Lsh0x](https://github.com/Lsh0x))
-- feat: updates to 2022 [\#1](https://github.com/Lsh0x/tars/pull/1) ([Lsh0x](https://github.com/Lsh0x))
+## [0.0.2](https://github.com/Lsh0x/rsta/releases/tag/v0.0.2)
 
+[Full Changelog](https://github.com/Lsh0x/rsta/compare/v0.0.1...v0.0.2)
 
+**Changes:**
 
-\* *This Changelog was automatically generated by [github_changelog_generator](https://github.com/github-changelog-generator/github-changelog-generator)*
+- Updated library version
+- Fixed class naming inconsistencies
+- Improved code examples in documentation
+
+## [0.0.1](https://github.com/Lsh0x/rsta/releases/tag/v0.0.1)
+
+**Changes:**
+
+- Initial release of RSTA library
+- Implemented basic technical indicators:
+  - Trend indicators (SMA, EMA)
+  - Momentum indicators (RSI, Stochastic Oscillator, Williams %R)
+  - Volume indicators (OBV, Volume Rate of Change, Accumulation/Distribution Line)
+  - Volatility indicators (ATR, Bollinger Bands, Keltner Channels)
+- Added comprehensive documentation and examples
+
+\* *This Changelog was automatically generated and then customized for RSTA*
