--- src/indicators/momentum.rs
@@ -1,853 +1,0 @@
-//! Momentum indicators
-//!
-//! This module contains momentum indicators like RSI, Stochastic Oscillator, and Williams %R.
-
-use crate::indicators::utils::{validate_data_length, validate_period};
-use crate::indicators::{Candle, Indicator, IndicatorError};
-use std::collections::VecDeque;
-
-/// Relative Strength Index (RSI) indicator
-///
-/// RSI measures the magnitude of recent price changes to evaluate
-/// overbought or oversold conditions. The RSI ranges from 0 to 100.
-/// Traditionally, RSI values of 70 or above indicate overbought conditions,
-/// while values of 30 or below indicate oversold conditions.
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::momentum::RelativeStrengthIndex;
-/// use rsta::indicators::Indicator;
-///
-/// // Create a 14-period RSI
-/// let mut rsi = RelativeStrengthIndex::new(14).unwrap();
-///
-/// // Price data
-/// let prices = vec![44.34, 44.09, 44.15, 43.61, 44.33, 44.83, 45.10, 45.42,
-///                   45.84, 46.08, 45.89, 46.03, 45.61, 46.28, 46.28, 46.00,
-///                   46.03, 46.41, 46.22, 45.64, 46.21, 46.25, 45.71, 46.45];
-///
-/// // Calculate RSI values
-/// let rsi_values = rsi.calculate(&prices).unwrap();
-/// ```
-#[derive(Debug)]
-pub struct RelativeStrengthIndex {
-    period: usize,
-    prev_price: Option<f64>,
-    gains: VecDeque<f64>,
-    losses: VecDeque<f64>,
-    avg_gain: Option<f64>,
-    avg_loss: Option<f64>,
-}
-
-impl RelativeStrengthIndex {
-    /// Create a new RelativeStrengthIndex indicator
-    ///
-    /// # Arguments
-    /// * `period` - The period for RSI calculation (must be at least 1)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new RelativeStrengthIndex or an error
-    pub fn new(period: usize) -> Result<Self, IndicatorError> {
-        validate_period(period, 1)?;
-
-        Ok(Self {
-            period,
-            prev_price: None,
-            gains: VecDeque::with_capacity(period),
-            losses: VecDeque::with_capacity(period),
-            avg_gain: None,
-            avg_loss: None,
-        })
-    }
-
-    /// Calculate a single RSI value from average gain and loss
-    ///
-    /// # Arguments
-    /// * `avg_gain` - The average gain over the period
-    /// * `avg_loss` - The average loss over the period
-    ///
-    /// # Returns
-    /// * `f64` - The RSI value
-    fn calculate_rsi(avg_gain: f64, avg_loss: f64) -> f64 {
-        if avg_loss == 0.0 {
-            return 100.0;
-        }
-
-        let rs = avg_gain / avg_loss;
-        100.0 - (100.0 / (1.0 + rs))
-    }
-}
-
-impl Indicator<f64, f64> for RelativeStrengthIndex {
-    fn calculate(&mut self, data: &[f64]) -> Result<Vec<f64>, IndicatorError> {
-        validate_data_length(data, self.period + 1)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.period);
-
-        // Reset state
-        self.reset();
-
-        // First, calculate price changes
-        let mut price_changes = Vec::with_capacity(n - 1);
-        for i in 1..n {
-            price_changes.push(data[i] - data[i - 1]);
-        }
-
-        // Calculate initial gains and losses
-        let mut gains = Vec::with_capacity(self.period);
-        let mut losses = Vec::with_capacity(self.period);
-
-        for &change in price_changes.iter().take(self.period) {
-            if change > 0.0 {
-                gains.push(change);
-                losses.push(0.0);
-            } else {
-                gains.push(0.0);
-                losses.push(-change);
-            }
-        }
-
-        // Calculate first average gain and loss
-        let mut avg_gain = gains.iter().sum::<f64>() / self.period as f64;
-        let mut avg_loss = losses.iter().sum::<f64>() / self.period as f64;
-
-        // Calculate first RSI
-        result.push(Self::calculate_rsi(avg_gain, avg_loss));
-
-        // Calculate the rest using the smoothed method
-        for change in price_changes.iter().skip(self.period).copied() {
-            let gain = if change > 0.0 { change } else { 0.0 };
-            let loss = if change < 0.0 { -change } else { 0.0 };
-
-            // Use Wilder's smoothing method
-            avg_gain = (avg_gain * (self.period - 1) as f64 + gain) / self.period as f64;
-            avg_loss = (avg_loss * (self.period - 1) as f64 + loss) / self.period as f64;
-
-            result.push(Self::calculate_rsi(avg_gain, avg_loss));
-        }
-
-        Ok(result)
-    }
-
-    fn next(&mut self, value: f64) -> Result<Option<f64>, IndicatorError> {
-        if let Some(prev) = self.prev_price {
-            let change = value - prev;
-            let gain = if change > 0.0 { change } else { 0.0 };
-            let loss = if change < 0.0 { -change } else { 0.0 };
-
-            self.gains.push_back(gain);
-            self.losses.push_back(loss);
-
-            if self.gains.len() > self.period {
-                self.gains.pop_front();
-                self.losses.pop_front();
-            }
-
-            if self.gains.len() < self.period {
-                self.avg_gain = None;
-                self.avg_loss = None;
-                self.prev_price = Some(value);
-                return Ok(None);
-            }
-
-            // Calculate/update average gain and loss
-            if let (Some(avg_gain), Some(avg_loss)) = (self.avg_gain, self.avg_loss) {
-                // Use Wilder's smoothing method for ongoing calculations
-                self.avg_gain =
-                    Some((avg_gain * (self.period - 1) as f64 + gain) / self.period as f64);
-                self.avg_loss =
-                    Some((avg_loss * (self.period - 1) as f64 + loss) / self.period as f64);
-            } else {
-                // Initial average calculation
-                self.avg_gain = Some(self.gains.iter().sum::<f64>() / self.period as f64);
-                self.avg_loss = Some(self.losses.iter().sum::<f64>() / self.period as f64);
-            }
-
-            let rsi = Self::calculate_rsi(self.avg_gain.unwrap(), self.avg_loss.unwrap());
-
-            self.prev_price = Some(value);
-            Ok(Some(rsi))
-        } else {
-            self.prev_price = Some(value);
-            Ok(None)
-        }
-    }
-
-    fn reset(&mut self) {
-        self.prev_price = None;
-        self.gains.clear();
-        self.losses.clear();
-        self.avg_gain = None;
-        self.avg_loss = None;
-    }
-}
-
-/// Stochastic Oscillator
-///
-/// The Stochastic Oscillator is a momentum indicator that shows the location of the close
-/// relative to the high-low range over a set number of periods.
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::momentum::{StochasticOscillator, StochasticResult};
-/// use rsta::indicators::{Indicator, Candle};
-///
-/// // Create a Stochastic Oscillator with %K period of 14 and %D period of 3
-/// let mut stoch = StochasticOscillator::new(14, 3).unwrap();
-///
-/// // Create price data with OHLC values (need at least 16 candles for 14 %K period + 3 %D period - 1)
-/// let candles = vec![
-///     // Initial candles for the calculation window
-///     Candle { timestamp: 1, open: 42.0, high: 43.0, low: 41.0, close: 42.5, volume: 1000.0 },
-///     Candle { timestamp: 2, open: 42.5, high: 43.5, low: 41.5, close: 43.0, volume: 1100.0 },
-///     Candle { timestamp: 3, open: 43.0, high: 44.0, low: 42.0, close: 43.5, volume: 1200.0 },
-///     Candle { timestamp: 4, open: 43.5, high: 44.5, low: 42.5, close: 44.0, volume: 1300.0 },
-///     Candle { timestamp: 5, open: 44.0, high: 45.0, low: 43.0, close: 44.5, volume: 1400.0 },
-///     Candle { timestamp: 6, open: 44.5, high: 45.5, low: 43.5, close: 45.0, volume: 1500.0 },
-///     Candle { timestamp: 7, open: 45.0, high: 46.0, low: 44.0, close: 45.5, volume: 1600.0 },
-///     Candle { timestamp: 8, open: 45.5, high: 46.5, low: 44.5, close: 46.0, volume: 1700.0 },
-///     Candle { timestamp: 9, open: 46.0, high: 47.0, low: 45.0, close: 46.5, volume: 1800.0 },
-///     Candle { timestamp: 10, open: 46.5, high: 47.5, low: 45.5, close: 47.0, volume: 1900.0 },
-///     Candle { timestamp: 11, open: 47.0, high: 48.0, low: 46.0, close: 47.5, volume: 2000.0 },
-///     Candle { timestamp: 12, open: 47.5, high: 48.5, low: 46.5, close: 48.0, volume: 2100.0 },
-///     Candle { timestamp: 13, open: 48.0, high: 49.0, low: 47.0, close: 48.5, volume: 2200.0 },
-///     Candle { timestamp: 14, open: 48.5, high: 49.5, low: 47.5, close: 49.0, volume: 2300.0 },
-///     // Additional candles to calculate %D (3-period SMA of %K)
-///     Candle { timestamp: 15, open: 49.0, high: 50.0, low: 48.0, close: 49.5, volume: 2400.0 },
-///     Candle { timestamp: 16, open: 49.5, high: 50.5, low: 48.5, close: 50.0, volume: 2500.0 },
-///     Candle { timestamp: 17, open: 50.0, high: 51.0, low: 49.0, close: 49.0, volume: 2600.0 }, // Price drop
-/// ];
-///
-/// // Calculate Stochastic values
-/// match stoch.calculate(&candles) {
-///     Ok(stoch_values) => {
-///         // Access the latest Stochastic values
-///         if let Some(latest) = stoch_values.last() {
-///             println!("%K (Fast): {:.2}", latest.k); // Example output: %K (Fast): 50.00
-///             println!("%D (Slow): {:.2}", latest.d); // Example output: %D (Slow): 66.67
-///             
-///             // Interpret the values
-///             if latest.k > 80.0 && latest.d > 80.0 {
-///                 println!("Overbought condition");
-///             } else if latest.k < 20.0 && latest.d < 20.0 {
-///                 println!("Oversold condition");
-///             }
-///             
-///             // Check for crossovers
-///             if stoch_values.len() >= 2 {
-///                 let previous = stoch_values[stoch_values.len() - 2];
-///                 if latest.k > latest.d && previous.k <= previous.d {
-///                     println!("Bullish crossover: %K crossed above %D");
-///                 } else if latest.k < latest.d && previous.k >= previous.d {
-///                     println!("Bearish crossover: %K crossed below %D");
-///                 }
-///             }
-///         }
-///     },
-///     Err(e) => {
-///         eprintln!("Error calculating Stochastic: {}", e);
-///     }
-/// }
-/// ```
-#[derive(Debug)]
-pub struct StochasticOscillator {
-    k_period: usize,
-    d_period: usize,
-    k_buffer: VecDeque<f64>,
-}
-
-impl StochasticOscillator {
-    /// Create a new StochasticOscillator
-    ///
-    /// # Arguments
-    /// * `k_period` - The %K period (typically 14) - must be at least 1
-    /// * `d_period` - The %D period (typically 3) - must be at least 1
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new StochasticOscillator or an error
-    pub fn new(k_period: usize, d_period: usize) -> Result<Self, IndicatorError> {
-        validate_period(k_period, 1)?;
-        validate_period(d_period, 1)?;
-
-        Ok(Self {
-            k_period,
-            d_period,
-            k_buffer: VecDeque::with_capacity(d_period),
-        })
-    }
-
-    /// Calculate %K value for a given candle
-    ///
-    /// # Arguments
-    /// * `candles` - The slice of candles to calculate %K from
-    /// * `idx` - The index of the current candle
-    ///
-    /// # Returns
-    /// * `f64` - The %K value
-    // StochasticOscillator fix - line 291
-    fn calculate_k(candles: &[Candle], idx: usize, period: usize) -> f64 {
-        if idx < period - 1 {
-            return 50.0; // Not enough data, return middle value
-        }
-
-        let current_close = candles[idx].close;
-
-        // Safe start index calculation to avoid integer underflow
-        let start_idx = idx.saturating_sub(period - 1);
-        let mut lowest_low = candles[start_idx].low;
-        let mut highest_high = candles[start_idx].high;
-
-        for candle in candles.iter().take(idx + 1).skip(start_idx + 1) {
-            lowest_low = lowest_low.min(candle.low);
-            highest_high = highest_high.max(candle.high);
-        }
-
-        if highest_high == lowest_low {
-            return 50.0; // Default to middle value when range is zero
-        }
-
-        ((current_close - lowest_low) / (highest_high - lowest_low)) * 100.0
-    }
-}
-
-/// Stochastic indicator result
-#[derive(Debug, Clone, Copy)]
-pub struct StochasticResult {
-    /// %K value (fast stochastic)
-    pub k: f64,
-    /// %D value (slow stochastic - SMA of %K)
-    pub d: f64,
-}
-
-impl Indicator<Candle, StochasticResult> for StochasticOscillator {
-    fn calculate(&mut self, data: &[Candle]) -> Result<Vec<StochasticResult>, IndicatorError> {
-        validate_data_length(data, self.k_period + self.d_period - 1)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.k_period - self.d_period + 2);
-
-        // Reset state
-        self.reset();
-
-        // Calculate %K values
-        let mut k_values = Vec::with_capacity(n);
-        for i in 0..n {
-            k_values.push(Self::calculate_k(data, i, self.k_period));
-        }
-
-        // We can only start calculating %D once we have k_period values
-        // We can only start calculating %D once we have k_period values
-        let k_start_idx = self.k_period - 1;
-        for (i, &k_value) in k_values.iter().enumerate().skip(k_start_idx) {
-            // Add to buffer
-            self.k_buffer.push_back(k_value);
-
-            if self.k_buffer.len() > self.d_period {
-                self.k_buffer.pop_front();
-            }
-
-            if self.k_buffer.len() == self.d_period {
-                // Calculate %D (SMA of %K)
-                let d = self.k_buffer.iter().sum::<f64>() / self.d_period as f64;
-
-                result.push(StochasticResult { k: k_values[i], d });
-            }
-        }
-        Ok(result)
-    }
-
-    fn next(&mut self, _value: Candle) -> Result<Option<StochasticResult>, IndicatorError> {
-        // Implementation would require storing the last k_period candles
-        // For simplicity, we're leaving this as an exercise
-        Err(IndicatorError::CalculationError(
-            "Real-time calculation of Stochastic Oscillator requires storing previous candles"
-                .to_string(),
-        ))
-    }
-
-    fn reset(&mut self) {
-        self.k_buffer.clear();
-    }
-}
-
-/// Williams %R
-///
-/// Williams %R is a momentum indicator that is the inverse of the Fast Stochastic Oscillator.
-/// It reflects the level of the close relative to the highest high for the look-back period.
-///
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::momentum::WilliamsR;
-/// use rsta::indicators::{Indicator, Candle};
-///
-/// // Create a 14-period Williams %R
-/// let mut williams_r = WilliamsR::new(14).unwrap();
-///
-/// // Price data as candles - Need at least 14 data points
-/// let candles = vec![
-///     // Initial candles for the calculation window
-///     Candle { timestamp: 1, open: 42.0, high: 43.0, low: 41.0, close: 42.5, volume: 1000.0 },
-///     Candle { timestamp: 2, open: 42.5, high: 43.5, low: 41.5, close: 43.0, volume: 1100.0 },
-///     Candle { timestamp: 3, open: 43.0, high: 44.0, low: 42.0, close: 43.5, volume: 1200.0 },
-///     Candle { timestamp: 4, open: 43.5, high: 44.5, low: 42.5, close: 44.0, volume: 1300.0 },
-///     Candle { timestamp: 5, open: 44.0, high: 45.0, low: 43.0, close: 44.5, volume: 1400.0 },
-///     Candle { timestamp: 6, open: 44.5, high: 45.5, low: 43.5, close: 45.0, volume: 1500.0 },
-///     Candle { timestamp: 7, open: 45.0, high: 46.0, low: 44.0, close: 45.5, volume: 1600.0 },
-///     Candle { timestamp: 8, open: 45.5, high: 46.5, low: 44.5, close: 46.0, volume: 1700.0 },
-///     Candle { timestamp: 9, open: 46.0, high: 47.0, low: 45.0, close: 46.5, volume: 1800.0 },
-///     Candle { timestamp: 10, open: 46.5, high: 47.5, low: 45.5, close: 47.0, volume: 1900.0 },
-///     Candle { timestamp: 11, open: 47.0, high: 48.0, low: 46.0, close: 47.5, volume: 2000.0 },
-///     Candle { timestamp: 12, open: 47.5, high: 48.5, low: 46.5, close: 48.0, volume: 2100.0 },
-///     Candle { timestamp: 13, open: 48.0, high: 49.0, low: 47.0, close: 48.5, volume: 2200.0 },
-///     Candle { timestamp: 14, open: 48.5, high: 49.5, low: 47.5, close: 49.0, volume: 2300.0 },
-///     // Last candle for testing decreasing price
-///     Candle { timestamp: 15, open: 49.0, high: 49.5, low: 47.0, close: 47.5, volume: 2400.0 },
-/// ];
-///
-/// // Calculate Williams %R values with error handling
-/// match williams_r.calculate(&candles) {
-///     Ok(r_values) => {
-///         // Access the latest value
-///         if let Some(latest_r) = r_values.last() {
-///             println!("Williams %R: {:.2}", latest_r); // Example output: Williams %R: -80.00
-///             
-///             // Interpret the Williams %R value
-///             // Note: Williams %R ranges from -100 to 0
-///             if *latest_r > -20.0 {
-///                 println!("Overbought condition (> -20)");
-///             } else if *latest_r < -80.0 {
-///                 println!("Oversold condition (< -80)");
-///             } else {
-///                 println!("Neutral territory");
-///             }
-///             
-///             // Check for momentum shifts
-///             if r_values.len() >= 2 {
-///                 let previous_r = r_values[r_values.len() - 2];
-///                 if *latest_r > previous_r {
-///                     println!("Momentum improving (value rising)");
-///                 } else if *latest_r < previous_r {
-///                     println!("Momentum declining (value falling)");
-///                 }
-///             }
-///         }
-///     },
-///     Err(e) => {
-///         eprintln!("Error calculating Williams %R: {}", e);
-///     }
-/// }
-/// ```
-pub struct WilliamsR {
-    period: usize,
-}
-
-impl WilliamsR {
-    /// Create a new WilliamsR indicator
-    ///
-    /// # Arguments
-    /// * `period` - The period for Williams %R calculation (must be at least 1)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new WilliamsR indicator or an error
-    pub fn new(period: usize) -> Result<Self, IndicatorError> {
-        validate_period(period, 1)?;
-
-        Ok(Self { period })
-    }
-
-    /// Calculate Williams %R value for a given candle range
-    ///
-    /// # Arguments
-    /// * `candles` - The slice of candles to calculate %R from
-    /// * `idx` - The index of the current candle
-    /// * `period` - The period for Williams %R calculation
-    ///
-    /// # Returns
-    /// * `f64` - The Williams %R value
-    fn calculate_r(candles: &[Candle], idx: usize, period: usize) -> f64 {
-        if idx < period - 1 {
-            return -50.0; // Not enough data, return middle value
-        }
-
-        let current_close = candles[idx].close;
-
-        // Safe start index calculation to avoid integer underflow
-        let start_idx = idx.saturating_sub(period - 1);
-        let mut lowest_low = candles[start_idx].low;
-        let mut highest_high = candles[start_idx].high;
-
-        for candle in candles.iter().take(idx + 1).skip(start_idx + 1) {
-            lowest_low = lowest_low.min(candle.low);
-            highest_high = highest_high.max(candle.high);
-        }
-
-        if highest_high == lowest_low {
-            return -50.0; // Default to middle value when range is zero
-        }
-
-        // Williams %R formula: ((Highest High - Close) / (Highest High - Lowest Low)) * -100
-        ((highest_high - current_close) / (highest_high - lowest_low)) * -100.0
-    }
-}
-
-impl Indicator<Candle, f64> for WilliamsR {
-    fn calculate(&mut self, data: &[Candle]) -> Result<Vec<f64>, IndicatorError> {
-        validate_data_length(data, self.period)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.period + 1);
-
-        // Calculate Williams %R for each period
-        for i in (self.period - 1)..n {
-            let r_value = Self::calculate_r(data, i, self.period);
-            result.push(r_value);
-        }
-
-        Ok(result)
-    }
-
-    fn next(&mut self, _value: Candle) -> Result<Option<f64>, IndicatorError> {
-        // This implementation would require storing the last period candles
-        // For simplicity, we'll return an error message that real-time calculation
-        // requires previous candle storage
-        Err(IndicatorError::CalculationError(
-            "Real-time calculation of Williams %R requires storing previous candles".to_string(),
-        ))
-    }
-
-    fn reset(&mut self) {
-        // No state to reset in this basic implementation
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::indicators::Candle;
-
-    // RelativeStrengthIndex Tests
-    #[test]
-    fn test_rsi_new() {
-        // Valid period should work
-        assert!(RelativeStrengthIndex::new(14).is_ok());
-
-        // Invalid period should fail
-        assert!(RelativeStrengthIndex::new(0).is_err());
-    }
-
-    #[test]
-    fn test_rsi_calculation() {
-        let mut rsi = RelativeStrengthIndex::new(3).unwrap();
-
-        // Sample price data
-        let prices = vec![10.0, 11.0, 10.5, 11.5, 12.0, 11.0, 11.5];
-
-        let result = rsi.calculate(&prices).unwrap();
-        assert_eq!(result.len(), 4); // 7 prices - 3 period = 4 results
-
-        // First RSI: price changes = [1.0, -0.5, 1.0]
-        // Average gain = (1.0 + 0.0 + 1.0) / 3 = 0.6667
-        // Average loss = (0.0 + 0.5 + 0.0) / 3 = 0.1667
-        // RS = 0.6667 / 0.1667 = 4.0
-        // RSI = 100 - (100 / (1 + 4.0)) = 100 - 20 = 80.0
-        assert!((result[0] - 80.0).abs() < 0.01);
-
-        // Check final value is correct
-        // Last price change is 0.5, a gain
-        // Using Wilder's smoothing method
-        let last_value = result.last().unwrap();
-        assert!(*last_value >= 0.0 && *last_value <= 100.0);
-    }
-
-    #[test]
-    fn test_rsi_next() {
-        let mut rsi = RelativeStrengthIndex::new(3).unwrap();
-
-        // Initial values - not enough data yet
-        assert_eq!(rsi.next(10.0).unwrap(), None);
-        assert_eq!(rsi.next(11.0).unwrap(), None);
-        assert_eq!(rsi.next(10.5).unwrap(), None);
-
-        // Fourth value - now we have RSI
-        let first_rsi = rsi.next(11.5).unwrap();
-        assert!(first_rsi.is_some());
-        let first_rsi_value = first_rsi.unwrap();
-        assert!((0.0..=100.0).contains(&first_rsi_value));
-
-        // More values - should keep producing results
-        assert!(rsi.next(12.0).unwrap().is_some());
-        assert!(rsi.next(11.0).unwrap().is_some());
-    }
-
-    #[test]
-    fn test_rsi_reset() {
-        let mut rsi = RelativeStrengthIndex::new(3).unwrap();
-
-        // Add some values
-        rsi.next(10.0).unwrap();
-        rsi.next(11.0).unwrap();
-        rsi.next(10.5).unwrap();
-        rsi.next(11.5).unwrap(); // This should produce a result
-
-        // Reset
-        rsi.reset();
-
-        // Should be back to initial state
-        assert_eq!(rsi.next(12.0).unwrap(), None);
-    }
-
-    // StochasticOscillator Tests
-    #[test]
-    fn test_stochastic_new() {
-        // Valid periods should work
-        assert!(StochasticOscillator::new(14, 3).is_ok());
-
-        // Invalid periods should fail
-        assert!(StochasticOscillator::new(0, 3).is_err());
-        assert!(StochasticOscillator::new(14, 0).is_err());
-    }
-
-    #[test]
-    fn test_stochastic_calculation() {
-        let mut stoch = StochasticOscillator::new(3, 2).unwrap();
-
-        // Create a series of candles with predictable pattern
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 12.0,
-                low: 9.0,
-                close: 11.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 11.0,
-                high: 13.0,
-                low: 10.0,
-                close: 12.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 3,
-                open: 12.0,
-                high: 14.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 4,
-                open: 13.0,
-                high: 15.0,
-                low: 12.0,
-                close: 14.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 5,
-                open: 14.0,
-                high: 16.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-        ];
-
-        let result = stoch.calculate(&candles).unwrap();
-
-        // We expect: 3 candles for k_period + 2 candles for d_period - 1 = 4 candles needed
-        // So we get 5 - 4 + 1 = 2 results
-        assert_eq!(result.len(), 2);
-
-        // Verify the StochasticResult struct has valid values
-        for stoch_result in &result {
-            assert!(stoch_result.k >= 0.0 && stoch_result.k <= 100.0);
-            assert!(stoch_result.d >= 0.0 && stoch_result.d <= 100.0);
-        }
-    }
-
-    #[test]
-    fn test_stochastic_next_error() {
-        let mut stoch = StochasticOscillator::new(14, 3).unwrap();
-        let candle = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-
-        // The next method should return an error as noted in the implementation
-        assert!(stoch.next(candle).is_err());
-    }
-
-    #[test]
-    fn test_stochastic_reset() {
-        let mut stoch = StochasticOscillator::new(14, 3).unwrap();
-
-        // Create data
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 12.0,
-                low: 9.0,
-                close: 11.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 11.0,
-                high: 13.0,
-                low: 10.0,
-                close: 12.0,
-                volume: 1000.0,
-            },
-        ];
-
-        // Calculate something
-        let _ = stoch.calculate(&candles);
-
-        // Reset
-        stoch.reset();
-
-        // k_buffer should be empty after reset
-        // We can't directly test the internal state, but we can test the behavior
-        // by doing a calculation that requires an empty state
-    }
-
-    // WilliamsR Tests
-    #[test]
-    fn test_williams_r_new() {
-        // Valid period should work
-        assert!(WilliamsR::new(14).is_ok());
-
-        // Invalid period should fail
-        assert!(WilliamsR::new(0).is_err());
-    }
-
-    #[test]
-    fn test_williams_r_calculation() {
-        let mut williams_r = WilliamsR::new(3).unwrap();
-
-        // Create a series of candles with predictable pattern
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 12.0,
-                low: 9.0,
-                close: 11.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 11.0,
-                high: 13.0,
-                low: 10.0,
-                close: 12.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 3,
-                open: 12.0,
-                high: 14.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 4,
-                open: 13.0,
-                high: 15.0,
-                low: 12.0,
-                close: 14.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 5,
-                open: 14.0,
-                high: 16.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-        ];
-
-        let result = williams_r.calculate(&candles).unwrap();
-
-        // We expect 5 - 3 + 1 = 3 results
-        assert_eq!(result.len(), 3);
-
-        // Williams %R should be between -100 and 0
-        for r_value in &result {
-            assert!(*r_value <= 0.0 && *r_value >= -100.0);
-        }
-
-        // For the 3rd candle:
-        // High = 14.0, Low = 11.0, Close = 13.0
-        // %R calculation may vary significantly due to implementation details
-        assert!(result[0] <= 0.0 && result[0] >= -100.0); // Just verify it's in valid range
-    }
-
-    #[test]
-    fn test_williams_r_next_error() {
-        let mut williams_r = WilliamsR::new(14).unwrap();
-        let candle = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-
-        // The next method should return an error as noted in the implementation
-        assert!(williams_r.next(candle).is_err());
-    }
-
-    #[test]
-    fn test_calculate_r() {
-        // Test the internal calculate_r function with a simple case
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 10.0,
-                low: 5.0,
-                close: 7.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 7.0,
-                high: 15.0,
-                low: 7.0,
-                close: 8.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 3,
-                open: 8.0,
-                high: 10.0,
-                low: 6.0,
-                close: 7.0,
-                volume: 1000.0,
-            },
-        ];
-
-        // For period 3, idx 2:
-        // Highest high = 15.0, Lowest low = 5.0, Close = 7.0
-        // %R = ((15 - 7) / (15 - 5)) * -100 = -80.0
-        let r_value = WilliamsR::calculate_r(&candles, 2, 3);
-        assert!((r_value - (-80.0)).abs() < 0.01);
-    }
-}
