+++ src/indicators/momentum/mod.rs
@@ -1,0 +1,16 @@
+//! Momentum indicators
+//!
+//! This module contains indicators that measure the rate of change or momentum of price movements.
+//! These include the Relative Strength Index (RSI), Stochastic Oscillator, and Williams %R.
+//!
+//! Momentum indicators are useful for identifying overbought and oversold conditions,
+//! trend strength, and potential reversals.
+
+pub mod rsi;
+pub mod stochastic_oscillator;
+pub mod williams_r;
+
+// Re-export public types to maintain the same interface
+pub use self::rsi::Rsi;
+pub use self::stochastic_oscillator::{StochasticOscillator, StochasticResult};
+pub use self::williams_r::WilliamsR;
