+++ src/indicators/trend/mod.rs
@@ -1,0 +1,5 @@
+pub mod ema;
+pub mod sma;
+
+pub use self::ema::Ema;
+pub use self::sma::Sma;
