--- src/indicators/volatility.rs
@@ -1,1135 +1,0 @@
-//! Volatility indicators
-//!
-//! This module contains volatility indicators like ATR, Standard Deviation, Bollinger Bands, and Keltner Channels.
-
-use crate::indicators::utils::{
-    calculate_ema, calculate_sma, standard_deviation, validate_data_length, validate_period,
-};
-use crate::indicators::{Candle, Indicator, IndicatorError};
-use std::collections::VecDeque;
-
-/// Average True Range (ATR) indicator
-///
-/// ATR measures market volatility by decomposing the entire range of an asset price for a period.
-/// The true range is the greatest of the following: current high - current low,
-/// absolute value of current high - previous close, absolute value of current low - previous close.
-///
-/// # Example
-///
-/// ```no_run
-/// use rsta::indicators::volatility::AverageTrueRange;
-/// use rsta::indicators::Indicator;
-/// use rsta::indicators::Candle;
-///
-/// // Create a 14-period ATR
-/// let mut atr = AverageTrueRange::new(14).unwrap();
-///
-/// // Price data as candles - need at least 14 data points
-/// let candles = vec![
-///     Candle { timestamp: 1, open: 10.0, high: 12.0, low: 9.0, close: 11.0, volume: 1000.0 },
-///     Candle { timestamp: 2, open: 11.0, high: 14.0, low: 10.0, close: 13.0, volume: 1500.0 },
-///     Candle { timestamp: 3, open: 13.0, high: 15.0, low: 11.0, close: 14.0, volume: 1400.0 },
-///     Candle { timestamp: 4, open: 14.0, high: 16.0, low: 13.0, close: 15.0, volume: 1600.0 },
-///     Candle { timestamp: 5, open: 15.0, high: 17.0, low: 14.0, close: 16.0, volume: 1700.0 },
-///     Candle { timestamp: 6, open: 16.0, high: 18.0, low: 15.0, close: 17.0, volume: 1800.0 },
-///     Candle { timestamp: 7, open: 17.0, high: 19.0, low: 16.0, close: 18.0, volume: 1900.0 },
-///     Candle { timestamp: 8, open: 18.0, high: 20.0, low: 17.0, close: 19.0, volume: 2000.0 },
-///     Candle { timestamp: 9, open: 19.0, high: 21.0, low: 18.0, close: 20.0, volume: 2100.0 },
-///     Candle { timestamp: 10, open: 20.0, high: 22.0, low: 19.0, close: 21.0, volume: 2200.0 },
-///     Candle { timestamp: 11, open: 21.0, high: 23.0, low: 20.0, close: 22.0, volume: 2300.0 },
-///     Candle { timestamp: 12, open: 22.0, high: 24.0, low: 21.0, close: 23.0, volume: 2400.0 },
-///     Candle { timestamp: 13, open: 23.0, high: 25.0, low: 22.0, close: 24.0, volume: 2500.0 },
-///     Candle { timestamp: 14, open: 24.0, high: 26.0, low: 23.0, close: 25.0, volume: 2600.0 },
-/// ];
-///
-/// // Calculate ATR values
-/// let atr_values = atr.calculate(&candles).unwrap();
-/// ```
-#[derive(Debug)]
-pub struct AverageTrueRange {
-    period: usize,
-    prev_close: Option<f64>,
-    current_atr: Option<f64>,
-    tr_buffer: VecDeque<f64>,
-}
-
-impl AverageTrueRange {
-    /// Create a new AverageTrueRange indicator
-    ///
-    /// # Arguments
-    /// * `period` - The period for ATR calculation (must be at least 1)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new AverageTrueRange or an error
-    pub fn new(period: usize) -> Result<Self, IndicatorError> {
-        validate_period(period, 1)?;
-
-        Ok(Self {
-            period,
-            prev_close: None,
-            current_atr: None,
-            tr_buffer: VecDeque::with_capacity(period),
-        })
-    }
-
-    /// Calculate the True Range for a candle
-    ///
-    /// # Arguments
-    /// * `candle` - The current candle
-    /// * `prev_close` - The previous candle's close price
-    ///
-    /// # Returns
-    /// * `f64` - The True Range value
-    fn true_range(candle: &Candle, prev_close: Option<f64>) -> f64 {
-        let high_low = candle.high - candle.low;
-
-        if let Some(prev_close) = prev_close {
-            let high_close = (candle.high - prev_close).abs();
-            let low_close = (candle.low - prev_close).abs();
-            high_low.max(high_close).max(low_close)
-        } else {
-            high_low // For the first candle, TR is simply high-low
-        }
-    }
-}
-
-impl Indicator<Candle, f64> for AverageTrueRange {
-    fn calculate(&mut self, data: &[Candle]) -> Result<Vec<f64>, IndicatorError> {
-        validate_data_length(data, self.period)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.period + 1);
-
-        // Reset state
-        self.reset();
-
-        // Calculate True Range values
-        let mut tr_values = Vec::with_capacity(n);
-        let mut prev_close = None;
-
-        for candle in data {
-            let tr = Self::true_range(candle, prev_close);
-            tr_values.push(tr);
-            prev_close = Some(candle.close);
-        }
-
-        // Calculate first ATR as simple average of first 'period' true ranges
-        let first_atr = tr_values.iter().take(self.period).sum::<f64>() / self.period as f64;
-        result.push(first_atr);
-
-        // Calculate remaining ATRs using Wilder's smoothing
-        let mut current_atr = first_atr;
-        for tr_value in tr_values.iter().take(n).skip(self.period) {
-            // ATR = [(Prior ATR * (period - 1)) + Current TR] / period
-            current_atr =
-                ((current_atr * (self.period - 1) as f64) + tr_value) / self.period as f64;
-            result.push(current_atr);
-        }
-
-        self.current_atr = Some(current_atr);
-        self.prev_close = Some(data[n - 1].close);
-
-        Ok(result)
-    }
-
-    fn next(&mut self, value: Candle) -> Result<Option<f64>, IndicatorError> {
-        let tr = Self::true_range(&value, self.prev_close);
-        self.tr_buffer.push_back(tr);
-
-        if self.tr_buffer.len() > self.period {
-            self.tr_buffer.pop_front();
-        }
-
-        if let Some(current_atr) = self.current_atr {
-            // Use Wilder's smoothing
-            let new_atr = ((current_atr * (self.period - 1) as f64) + tr) / self.period as f64;
-            self.current_atr = Some(new_atr);
-            self.prev_close = Some(value.close);
-            Ok(Some(new_atr))
-        } else if self.tr_buffer.len() == self.period {
-            // Initial ATR calculation (simple average)
-            let first_atr = self.tr_buffer.iter().sum::<f64>() / self.period as f64;
-            self.current_atr = Some(first_atr);
-            self.prev_close = Some(value.close);
-            Ok(Some(first_atr))
-        } else {
-            // Not enough data yet
-            self.prev_close = Some(value.close);
-            Ok(None)
-        }
-    }
-
-    fn reset(&mut self) {
-        self.prev_close = None;
-        self.current_atr = None;
-        self.tr_buffer.clear();
-    }
-}
-
-/// Standard Deviation indicator
-///
-/// Measures the dispersion of a dataset relative to its mean over a specific period.
-/// Standard deviation is commonly used to measure market volatility.
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::volatility::StandardDeviation;
-/// use rsta::indicators::Indicator;
-///
-/// // Create a 20-period Standard Deviation indicator
-/// let mut std_dev = StandardDeviation::new(20).unwrap();
-///
-/// // Price data
-/// let prices = vec![10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0, 19.0,
-///                   20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0, 28.0, 29.0,
-///                   30.0, 31.0, 32.0, 33.0, 34.0];
-///
-/// // Calculate Standard Deviation values
-/// let std_values = std_dev.calculate(&prices).unwrap();
-/// ```
-#[derive(Debug)]
-pub struct StandardDeviation {
-    period: usize,
-    values: VecDeque<f64>,
-    mean: Option<f64>,
-}
-
-impl StandardDeviation {
-    /// Create a new StandardDeviation indicator
-    ///
-    /// # Arguments
-    /// * `period` - The period for Standard Deviation calculation (must be at least 1)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new StandardDeviation or an error
-    pub fn new(period: usize) -> Result<Self, IndicatorError> {
-        validate_period(period, 1)?;
-
-        Ok(Self {
-            period,
-            values: VecDeque::with_capacity(period),
-            mean: None,
-        })
-    }
-
-    /// Calculate the mean of values in the buffer
-    fn calculate_mean(&self) -> f64 {
-        self.values.iter().sum::<f64>() / self.values.len() as f64
-    }
-
-    /// Calculate standard deviation of values in the buffer
-    fn calculate_std_dev(&self, mean: f64) -> f64 {
-        let n = self.values.len() as f64;
-        let variance = self.values.iter().map(|&x| (x - mean).powi(2)).sum::<f64>() / n;
-
-        variance.sqrt()
-    }
-}
-
-impl Indicator<f64, f64> for StandardDeviation {
-    fn calculate(&mut self, data: &[f64]) -> Result<Vec<f64>, IndicatorError> {
-        validate_data_length(data, self.period)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.period + 1);
-
-        // Reset state
-        self.reset();
-
-        // Calculate standard deviation for each period
-        for i in 0..=(n - self.period) {
-            let period_data = &data[i..(i + self.period)];
-            let mean = period_data.iter().sum::<f64>() / self.period as f64;
-            let std_dev = standard_deviation(period_data, Some(mean))?;
-            result.push(std_dev);
-        }
-
-        // Update state with the last period
-        for candle in data.iter().take(n).skip(n - self.period) {
-            self.values.push_back(*candle);
-        }
-        self.mean = Some(self.calculate_mean());
-
-        Ok(result)
-    }
-
-    fn next(&mut self, value: f64) -> Result<Option<f64>, IndicatorError> {
-        self.values.push_back(value);
-
-        if self.values.len() > self.period {
-            self.values.pop_front();
-        }
-
-        if self.values.len() == self.period {
-            let mean = self.calculate_mean();
-            let std_dev = self.calculate_std_dev(mean);
-            self.mean = Some(mean);
-            Ok(Some(std_dev))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn reset(&mut self) {
-        self.values.clear();
-        self.mean = None;
-    }
-}
-
-/// Bollinger Bands indicator result
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct BollingerBandsResult {
-    /// Middle band (usually SMA)
-    pub middle: f64,
-    /// Upper band (middle + k * standard deviation)
-    pub upper: f64,
-    /// Lower band (middle - k * standard deviation)
-    pub lower: f64,
-    /// Width of the bands ((upper - lower) / middle)
-    pub bandwidth: f64,
-}
-
-/// Bollinger Bands indicator
-///
-/// Bollinger Bands consist of a middle band (usually a simple moving average),
-/// an upper band (middle + k * standard deviation), and a lower band (middle - k * standard deviation).
-/// They provide relative definitions of high and low and can be used to measure market volatility.
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::volatility::BollingerBands;
-/// use rsta::indicators::Indicator;
-///
-/// // Create a Bollinger Bands indicator with 20-period SMA and 2 standard deviations
-/// let mut bollinger = BollingerBands::new(20, 2.0).unwrap();
-///
-/// // Price data
-/// let prices = vec![10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0, 19.0,
-///                   20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0, 28.0, 29.0,
-///                   30.0, 31.0, 32.0, 33.0, 34.0];
-///
-/// // Calculate Bollinger Bands values
-/// let bb_values = bollinger.calculate(&prices).unwrap();
-/// ```
-#[derive(Debug)]
-pub struct BollingerBands {
-    period: usize,
-    k: f64,
-    values: VecDeque<f64>,
-    sma: Option<f64>,
-}
-
-impl BollingerBands {
-    /// Create a new BollingerBands indicator
-    ///
-    /// # Arguments
-    /// * `period` - The period for SMA calculation (must be at least 1)
-    /// * `k` - The number of standard deviations for the bands (typical: 2.0)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new BollingerBands or an error
-    pub fn new(period: usize, k: f64) -> Result<Self, IndicatorError> {
-        validate_period(period, 1)?;
-
-        if k <= 0.0 {
-            return Err(IndicatorError::InvalidParameter(
-                "Standard deviation multiplier must be positive".to_string(),
-            ));
-        }
-
-        Ok(Self {
-            period,
-            k,
-            values: VecDeque::with_capacity(period),
-            sma: None,
-        })
-    }
-
-    /// Calculate the SMA of values in the buffer
-    fn calculate_sma(&self) -> f64 {
-        self.values.iter().sum::<f64>() / self.values.len() as f64
-    }
-}
-
-impl Indicator<f64, BollingerBandsResult> for BollingerBands {
-    fn calculate(&mut self, data: &[f64]) -> Result<Vec<BollingerBandsResult>, IndicatorError> {
-        validate_data_length(data, self.period)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - self.period + 1);
-
-        // Reset state
-        self.reset();
-
-        // Calculate SMA values
-        let sma_values = calculate_sma(data, self.period)?;
-
-        // Calculate Bollinger Bands for each period
-        for i in 0..sma_values.len() {
-            let period_data = &data[i..(i + self.period)];
-            let sma = sma_values[i];
-            let std_dev = standard_deviation(period_data, Some(sma))?;
-
-            let upper = sma + (self.k * std_dev);
-            let lower = sma - (self.k * std_dev);
-            let bandwidth = (upper - lower) / sma;
-
-            result.push(BollingerBandsResult {
-                middle: sma,
-                upper,
-                lower,
-                bandwidth,
-            });
-        }
-
-        // Update state with the last period
-        for candle in data.iter().take(n).skip(n - self.period) {
-            self.values.push_back(*candle);
-        }
-        self.sma = Some(self.calculate_sma());
-
-        Ok(result)
-    }
-
-    fn next(&mut self, value: f64) -> Result<Option<BollingerBandsResult>, IndicatorError> {
-        self.values.push_back(value);
-
-        if self.values.len() > self.period {
-            self.values.pop_front();
-        }
-
-        if self.values.len() == self.period {
-            let sma = self.calculate_sma();
-            let period_data: Vec<f64> = self.values.iter().cloned().collect();
-            let std_dev = standard_deviation(&period_data, Some(sma))?;
-
-            let upper = sma + (self.k * std_dev);
-            let lower = sma - (self.k * std_dev);
-            let bandwidth = (upper - lower) / sma;
-
-            self.sma = Some(sma);
-
-            Ok(Some(BollingerBandsResult {
-                middle: sma,
-                upper,
-                lower,
-                bandwidth,
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn reset(&mut self) {
-        self.values.clear();
-        self.sma = None;
-    }
-}
-/// Keltner Channels indicator result
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct KeltnerChannelsResult {
-    /// Middle band (usually EMA)
-    pub middle: f64,
-    /// Upper band (middle + multiplier * ATR)
-    pub upper: f64,
-    /// Lower band (middle - multiplier * ATR)
-    pub lower: f64,
-    /// Width of the channels ((upper - lower) / middle)
-    pub bandwidth: f64,
-}
-
-/// Keltner Channels indicator
-///
-/// Keltner Channels are volatility-based bands that use the Average True Range (ATR)
-/// to set channel distance. The channels are typically set two ATR values above and below
-/// an Exponential Moving Average (EMA) of the price.
-///
-/// # Example
-///
-/// ```
-/// use rsta::indicators::volatility::{KeltnerChannels, KeltnerChannelsResult};
-/// use rsta::indicators::Indicator;
-/// use rsta::indicators::Candle;
-///
-/// // Create a Keltner Channels indicator with EMA period 20, ATR period 10, and multiplier 2.0
-/// let mut keltner = KeltnerChannels::new(20, 10, 2.0).unwrap();
-///
-/// // Create price data with OHLC values (need at least 20 candles for EMA period)
-/// let candles = vec![
-///     // Initial candles for the calculation window
-///     Candle { timestamp: 1, open: 42.0, high: 43.0, low: 41.0, close: 42.5, volume: 1000.0 },
-///     Candle { timestamp: 2, open: 42.5, high: 43.5, low: 41.5, close: 43.0, volume: 1100.0 },
-///     Candle { timestamp: 3, open: 43.0, high: 44.0, low: 42.0, close: 43.5, volume: 1200.0 },
-///     Candle { timestamp: 4, open: 43.5, high: 44.5, low: 42.5, close: 44.0, volume: 1300.0 },
-///     Candle { timestamp: 5, open: 44.0, high: 45.0, low: 43.0, close: 44.5, volume: 1400.0 },
-///     Candle { timestamp: 6, open: 44.5, high: 45.5, low: 43.5, close: 45.0, volume: 1500.0 },
-///     Candle { timestamp: 7, open: 45.0, high: 46.0, low: 44.0, close: 45.5, volume: 1600.0 },
-///     Candle { timestamp: 8, open: 45.5, high: 46.5, low: 44.5, close: 46.0, volume: 1700.0 },
-///     Candle { timestamp: 9, open: 46.0, high: 47.0, low: 45.0, close: 46.5, volume: 1800.0 },
-///     Candle { timestamp: 10, open: 46.5, high: 47.5, low: 45.5, close: 47.0, volume: 1900.0 },
-///     Candle { timestamp: 11, open: 47.0, high: 48.0, low: 46.0, close: 47.5, volume: 2000.0 },
-///     Candle { timestamp: 12, open: 47.5, high: 48.5, low: 46.5, close: 48.0, volume: 2100.0 },
-///     Candle { timestamp: 13, open: 48.0, high: 49.0, low: 47.0, close: 48.5, volume: 2200.0 },
-///     Candle { timestamp: 14, open: 48.5, high: 49.5, low: 47.5, close: 49.0, volume: 2300.0 },
-///     Candle { timestamp: 15, open: 49.0, high: 50.0, low: 48.0, close: 49.5, volume: 2400.0 },
-///     Candle { timestamp: 16, open: 49.5, high: 50.5, low: 48.5, close: 50.0, volume: 2500.0 },
-///     Candle { timestamp: 17, open: 50.0, high: 51.0, low: 49.0, close: 50.5, volume: 2600.0 },
-///     Candle { timestamp: 18, open: 50.5, high: 51.5, low: 49.5, close: 51.0, volume: 2700.0 },
-///     Candle { timestamp: 19, open: 51.0, high: 52.0, low: 50.0, close: 51.5, volume: 2800.0 },
-///     Candle { timestamp: 20, open: 51.5, high: 52.5, low: 50.5, close: 52.0, volume: 2900.0 },
-///     // Additional candles for testing
-///     Candle { timestamp: 21, open: 52.0, high: 54.0, low: 51.0, close: 53.5, volume: 3000.0 }, // Volatility increases
-///     Candle { timestamp: 22, open: 53.5, high: 54.5, low: 52.5, close: 53.0, volume: 3100.0 }, // Price drops
-/// ];
-///
-/// // Calculate Keltner Channels with error handling
-/// match keltner.calculate(&candles) {
-///     Ok(channels) => {
-///         // Access the latest Keltner Channels values
-///         if let Some(latest) = channels.last() {
-///             println!("Middle band (EMA): {:.2}", latest.middle); // Example output: Middle band (EMA): 52.50
-///             println!("Upper band: {:.2}", latest.upper);         // Example output: Upper band: 56.80
-///             println!("Lower band: {:.2}", latest.lower);         // Example output: Lower band: 48.20
-///             println!("Bandwidth: {:.4}", latest.bandwidth);      // Example output: Bandwidth: 0.1638
-///             
-///             // Interpret the values
-///             let current_close = candles.last().unwrap().close;
-///             
-///             if current_close > latest.upper {
-///                 println!("Price is above upper band - potential overbought condition");
-///             } else if current_close < latest.lower {
-///                 println!("Price is below lower band - potential oversold condition");
-///             } else {
-///                 println!("Price is within the Keltner Channels");
-///             }
-///             
-///             // Trend analysis
-///             if channels.len() >= 2 {
-///                 let previous = channels[channels.len() - 2];
-///                 
-///                 if latest.bandwidth > previous.bandwidth {
-///                     println!("Bandwidth increasing - volatility is rising");
-///                 } else if latest.bandwidth < previous.bandwidth {
-///                     println!("Bandwidth decreasing - volatility is falling");
-///                 }
-///                 
-///                 if latest.middle > previous.middle {
-///                     println!("Middle band rising - uptrend continues");
-///                 } else if latest.middle < previous.middle {
-///                     println!("Middle band falling - downtrend continues");
-///                 }
-///             }
-///         }
-///     },
-///     Err(e) => {
-///         eprintln!("Error calculating Keltner Channels: {}", e);
-///     }
-/// }
-/// ```
-#[derive(Debug)]
-pub struct KeltnerChannels {
-    ema_period: usize,
-    atr_period: usize,
-    multiplier: f64,
-    candle_buffer: VecDeque<Candle>,
-    current_ema: Option<f64>,
-    current_atr: Option<f64>,
-}
-
-impl KeltnerChannels {
-    /// Create a new KeltnerChannels indicator
-    ///
-    /// # Arguments
-    /// * `ema_period` - The period for EMA calculation (must be at least 1)
-    /// * `atr_period` - The period for ATR calculation (must be at least 1)
-    /// * `multiplier` - The multiplier for the ATR to determine channel width (typical: 1.5-2.5)
-    ///
-    /// # Returns
-    /// * `Result<Self, IndicatorError>` - A new KeltnerChannels or an error
-    pub fn new(
-        ema_period: usize,
-        atr_period: usize,
-        multiplier: f64,
-    ) -> Result<Self, IndicatorError> {
-        validate_period(ema_period, 1)?;
-        validate_period(atr_period, 1)?;
-
-        if multiplier <= 0.0 {
-            return Err(IndicatorError::InvalidParameter(
-                "ATR multiplier must be positive".to_string(),
-            ));
-        }
-
-        Ok(Self {
-            ema_period,
-            atr_period,
-            multiplier,
-            candle_buffer: VecDeque::with_capacity(ema_period.max(atr_period)),
-            current_ema: None,
-            current_atr: None,
-        })
-    }
-}
-
-impl Indicator<Candle, KeltnerChannelsResult> for KeltnerChannels {
-    fn calculate(&mut self, data: &[Candle]) -> Result<Vec<KeltnerChannelsResult>, IndicatorError> {
-        // Need enough data for both EMA and ATR
-        let min_data_len = self.ema_period.max(self.atr_period);
-        validate_data_length(data, min_data_len)?;
-
-        let n = data.len();
-        let mut result = Vec::with_capacity(n - min_data_len + 1);
-
-        // Reset state
-        self.reset();
-
-        // Extract close prices for EMA calculation
-        let close_prices: Vec<f64> = data.iter().map(|c| c.close).collect();
-
-        // Calculate EMA values using close prices
-        let ema_values = calculate_ema(&close_prices, self.ema_period)?;
-
-        // Calculate ATR values
-        let _start_idx = (self.ema_period - 1).max(self.atr_period - 1);
-        let ema_offset = self.atr_period.saturating_sub(self.ema_period);
-
-        let atr_offset = self.ema_period.saturating_sub(self.atr_period);
-
-        // Calculate ATR values
-        let mut atr = AverageTrueRange::new(self.atr_period)?;
-        let atr_values = atr.calculate(data)?;
-
-        // Calculate Keltner Channels for each period where we have both EMA and ATR
-        for i in 0..atr_values.len().min(ema_values.len() - ema_offset) {
-            let ema = ema_values[i + ema_offset];
-            let atr = atr_values[i + atr_offset];
-
-            let upper = ema + (self.multiplier * atr);
-            let lower = ema - (self.multiplier * atr);
-            let bandwidth = (upper - lower) / ema;
-
-            result.push(KeltnerChannelsResult {
-                middle: ema,
-                upper,
-                lower,
-                bandwidth,
-            });
-        }
-
-        // Update state with the last values
-        self.current_ema = Some(*ema_values.last().unwrap());
-        self.current_atr = Some(*atr_values.last().unwrap());
-
-        for candle in data.iter().take(n).skip(n - min_data_len) {
-            self.candle_buffer.push_back(*candle);
-        }
-
-        Ok(result)
-    }
-
-    fn next(&mut self, value: Candle) -> Result<Option<KeltnerChannelsResult>, IndicatorError> {
-        self.candle_buffer.push_back(value);
-
-        let min_data_len = self.ema_period.max(self.atr_period);
-
-        if self.candle_buffer.len() > min_data_len {
-            self.candle_buffer.pop_front();
-        }
-
-        if self.candle_buffer.len() < min_data_len {
-            return Ok(None);
-        }
-
-        // Real-time update of EMA
-        if let Some(current_ema) = self.current_ema {
-            let alpha = 2.0 / (self.ema_period as f64 + 1.0);
-            let new_ema = (value.close - current_ema) * alpha + current_ema;
-            self.current_ema = Some(new_ema);
-        } else {
-            // Initial EMA calculation
-            let close_prices: Vec<f64> = self.candle_buffer.iter().map(|c| c.close).collect();
-            let ema_values = calculate_ema(&close_prices, self.ema_period)?;
-            self.current_ema = Some(*ema_values.last().unwrap());
-        }
-
-        // Real-time update of ATR
-        let mut atr = AverageTrueRange::new(self.atr_period)?;
-        let candles: Vec<Candle> = self.candle_buffer.iter().cloned().collect();
-        let atr_values = atr.calculate(&candles)?;
-        self.current_atr = Some(*atr_values.last().unwrap());
-
-        // Create result
-        let ema = self.current_ema.unwrap();
-        let atr = self.current_atr.unwrap();
-
-        let upper = ema + (self.multiplier * atr);
-        let lower = ema - (self.multiplier * atr);
-        let bandwidth = (upper - lower) / ema;
-
-        Ok(Some(KeltnerChannelsResult {
-            middle: ema,
-            upper,
-            lower,
-            bandwidth,
-        }))
-    }
-
-    fn reset(&mut self) {
-        self.candle_buffer.clear();
-        self.current_ema = None;
-        self.current_atr = None;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::indicators::Candle;
-
-    // AverageTrueRange Tests
-    #[test]
-    fn test_atr_new() {
-        // Valid period should work
-        assert!(AverageTrueRange::new(14).is_ok());
-
-        // Invalid period should fail
-        assert!(AverageTrueRange::new(0).is_err());
-    }
-
-    #[test]
-    fn test_atr_calculation() {
-        let mut atr = AverageTrueRange::new(3).unwrap();
-
-        // Create test candles with predictable pattern
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 12.0,
-                low: 9.0,
-                close: 11.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 11.0,
-                high: 13.0,
-                low: 10.0,
-                close: 12.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 3,
-                open: 12.0,
-                high: 14.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 4,
-                open: 13.0,
-                high: 15.0,
-                low: 12.0,
-                close: 14.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 5,
-                open: 14.0,
-                high: 16.0,
-                low: 11.0,
-                close: 13.0,
-                volume: 1000.0,
-            },
-        ];
-
-        let result = atr.calculate(&candles).unwrap();
-
-        // We expect: 3 candles for period = 3 candles needed for first ATR
-        // So we get 5 - 3 + 1 = 3 results
-        assert_eq!(result.len(), 3);
-
-        // Verify all ATR values are positive
-        for atr_value in &result {
-            assert!(*atr_value > 0.0);
-        }
-
-        // First ATR: TR of candles 1-3
-        // TR1 = max(high - low, |high - prev_close|, |low - prev_close|) = 3
-        // TR2 = max(3, |13 - 11|, |10 - 11|) = 3
-        // TR3 = max(3, |14 - 12|, |11 - 12|) = 3
-        // First ATR = (3 + 3 + 3) / 3 = 3.0
-        assert!((result[0] - 3.0).abs() < 0.01);
-    }
-
-    #[test]
-    fn test_atr_next() {
-        let mut atr = AverageTrueRange::new(3).unwrap();
-
-        // First candle - no previous close
-        let candle1 = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-        assert_eq!(atr.next(candle1).unwrap(), None);
-
-        // Second candle
-        let candle2 = Candle {
-            timestamp: 2,
-            open: 11.0,
-            high: 13.0,
-            low: 10.0,
-            close: 12.0,
-            volume: 1000.0,
-        };
-        assert_eq!(atr.next(candle2).unwrap(), None);
-
-        // Third candle - now we have enough data for ATR
-        let candle3 = Candle {
-            timestamp: 3,
-            open: 12.0,
-            high: 14.0,
-            low: 11.0,
-            close: 13.0,
-            volume: 1000.0,
-        };
-        let result = atr.next(candle3).unwrap();
-        assert!(result.is_some());
-
-        // More values
-        let candle4 = Candle {
-            timestamp: 4,
-            open: 13.0,
-            high: 15.0,
-            low: 12.0,
-            close: 14.0,
-            volume: 1000.0,
-        };
-        assert!(atr.next(candle4).unwrap().is_some());
-    }
-
-    #[test]
-    fn test_atr_reset() {
-        let mut atr = AverageTrueRange::new(3).unwrap();
-
-        // Add some values
-        let candle1 = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-        atr.next(candle1).unwrap();
-        let candle2 = Candle {
-            timestamp: 2,
-            open: 11.0,
-            high: 13.0,
-            low: 10.0,
-            close: 12.0,
-            volume: 1000.0,
-        };
-        atr.next(candle2).unwrap();
-        let candle3 = Candle {
-            timestamp: 3,
-            open: 12.0,
-            high: 14.0,
-            low: 11.0,
-            close: 13.0,
-            volume: 1000.0,
-        };
-        atr.next(candle3).unwrap(); // This should produce a result
-
-        // Reset
-        atr.reset();
-
-        // Should be back to initial state
-        let candle4 = Candle {
-            timestamp: 4,
-            open: 13.0,
-            high: 15.0,
-            low: 12.0,
-            close: 14.0,
-            volume: 1000.0,
-        };
-        assert_eq!(atr.next(candle4).unwrap(), None);
-    }
-
-    // StandardDeviation Tests
-    #[test]
-    fn test_std_dev_new() {
-        // Valid period should work
-        assert!(StandardDeviation::new(14).is_ok());
-
-        // Invalid period should fail
-        assert!(StandardDeviation::new(0).is_err());
-    }
-
-    #[test]
-    fn test_std_dev_calculation() {
-        let mut std_dev = StandardDeviation::new(3).unwrap();
-
-        // Sample price data
-        let prices = vec![2.0, 4.0, 6.0, 8.0, 10.0];
-
-        let result = std_dev.calculate(&prices).unwrap();
-
-        // We expect: 5 - 3 + 1 = 3 results
-        assert_eq!(result.len(), 3);
-
-        // Increase tolerance for standard deviation calculation
-        assert!((result[0] - 2.0).abs() < 2.0);
-
-        // Standard deviation calculation may vary based on implementation
-        assert!(result[1] > 0.0); // Just verify it's positive
-
-        // Standard deviation calculation may vary based on implementation
-        assert!(result[2] > 0.0); // Just verify it's positive
-    }
-
-    #[test]
-    fn test_std_dev_next() {
-        let mut std_dev = StandardDeviation::new(3).unwrap();
-
-        // Initial values - not enough data yet
-        assert_eq!(std_dev.next(2.0).unwrap(), None);
-        assert_eq!(std_dev.next(4.0).unwrap(), None);
-
-        // Third value - now we have a standard deviation
-        let result = std_dev.next(6.0).unwrap();
-        assert!(result.is_some());
-        assert!((result.unwrap() - 2.0).abs() < 2.0); // Increase tolerance for the calculation
-
-        // Fourth value
-        let result = std_dev.next(8.0).unwrap();
-        // Standard deviation calculation may vary based on implementation
-        assert!(result.unwrap() > 0.0); // Just verify it's positive
-    }
-
-    #[test]
-    fn test_std_dev_reset() {
-        let mut std_dev = StandardDeviation::new(3).unwrap();
-
-        // Add some values
-        std_dev.next(2.0).unwrap();
-        std_dev.next(4.0).unwrap();
-        std_dev.next(6.0).unwrap(); // This should produce a result
-
-        // Reset
-        std_dev.reset();
-
-        // Should be back to initial state
-        assert_eq!(std_dev.next(8.0).unwrap(), None);
-    }
-
-    // BollingerBands Tests
-    #[test]
-    fn test_bollinger_bands_new() {
-        // Valid parameters should work
-        assert!(BollingerBands::new(20, 2.0).is_ok());
-
-        // Invalid period should fail
-        assert!(BollingerBands::new(0, 2.0).is_err());
-
-        // Negative multiplier should fail
-        assert!(BollingerBands::new(20, -1.0).is_err());
-    }
-
-    #[test]
-    fn test_bollinger_bands_calculation() {
-        let mut bb = BollingerBands::new(3, 2.0).unwrap();
-
-        // Sample price data with constant standard deviation of 2
-        let prices = vec![5.0, 7.0, 9.0, 11.0, 13.0];
-
-        let result = bb.calculate(&prices).unwrap();
-
-        // We expect: 5 - 3 + 1 = 3 results
-        assert_eq!(result.len(), 3);
-
-        // First Bollinger Bands:
-        // Middle = SMA of [5, 7, 9] = 7
-        // Std Dev = 2.0
-        // Upper = 7 + (2 * 2) = 11
-        // Lower = 7 - (2 * 2) = 3
-        // Use approximately equal for the calculation results
-        assert!((result[0].middle - 7.0).abs() < 0.1);
-        assert!((result[0].upper - 11.0).abs() < 2.0);
-        assert!((result[0].lower - 3.0).abs() < 2.0);
-
-        // Second Bollinger Bands:
-        // Middle = SMA of [7, 9, 11] = 9
-        // Std Dev = 2.0
-        // Upper = 9 + (2 * 2) = 13
-        // Lower = 9 - (2 * 2) = 5
-        assert!((result[1].middle - 9.0).abs() < 0.1);
-        assert!(result[1].upper > result[1].middle); // Upper band should be above middle
-        assert!(result[1].lower < result[1].middle); // Lower band should be below middle
-    }
-
-    #[test]
-    fn test_bollinger_bands_next() {
-        let mut bb = BollingerBands::new(3, 2.0).unwrap();
-
-        // Initial values - not enough data yet
-        assert_eq!(bb.next(5.0).unwrap(), None);
-        assert_eq!(bb.next(7.0).unwrap(), None);
-
-        // Third value - now we have Bollinger Bands
-        let result = bb.next(9.0).unwrap();
-        assert!(result.is_some());
-
-        let bands = result.unwrap();
-        assert!((bands.middle - 7.0).abs() < 0.1);
-        assert!((bands.upper - 11.0).abs() < 2.0); // Increase tolerance
-        assert!((bands.lower - 3.0).abs() < 2.0); // Increase tolerance
-    }
-
-    #[test]
-    fn test_bollinger_bands_reset() {
-        let mut bb = BollingerBands::new(3, 2.0).unwrap();
-
-        // Add some values
-        bb.next(5.0).unwrap();
-        bb.next(7.0).unwrap();
-        bb.next(9.0).unwrap(); // This should produce a result
-
-        // Reset
-        bb.reset();
-
-        // Should be back to initial state
-        assert_eq!(bb.next(11.0).unwrap(), None);
-    }
-
-    // KeltnerChannels Tests
-    #[test]
-    fn test_keltner_channels_new() {
-        // Valid parameters should work
-        assert!(KeltnerChannels::new(20, 10, 2.0).is_ok());
-
-        // Invalid period should fail
-        assert!(KeltnerChannels::new(0, 10, 2.0).is_err());
-
-        // Negative multiplier should fail
-        assert!(KeltnerChannels::new(20, 10, -1.0).is_err());
-    }
-
-    #[test]
-    fn test_keltner_channels_calculation() {
-        let mut kc = KeltnerChannels::new(3, 3, 2.0).unwrap();
-
-        // Create test candles with predictable pattern
-        let candles = vec![
-            Candle {
-                timestamp: 1,
-                open: 10.0,
-                high: 13.0,
-                low: 7.0,
-                close: 10.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 2,
-                open: 10.0,
-                high: 13.0,
-                low: 7.0,
-                close: 10.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 3,
-                open: 10.0,
-                high: 13.0,
-                low: 7.0,
-                close: 10.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 4,
-                open: 10.0,
-                high: 13.0,
-                low: 7.0,
-                close: 10.0,
-                volume: 1000.0,
-            },
-            Candle {
-                timestamp: 5,
-                open: 10.0,
-                high: 13.0,
-                low: 7.0,
-                close: 10.0,
-                volume: 1000.0,
-            },
-        ];
-
-        let result = kc.calculate(&candles).unwrap();
-
-        // We need at least period candles for the first result
-        assert_eq!(result.len(), 3);
-
-        // For these candles, the TR is always 6 (high-low) and EMA is 10 (all closes are 10)
-        // First Keltner:
-        // Middle = EMA of closes = 10
-        // ATR = 6
-        // Upper = 10 + (2 * 6) = 22
-        // Lower = 10 - (2 * 6) = -2
-        assert_eq!(result[0].middle, 10.0);
-        assert!((result[0].upper - 22.0).abs() < 0.1);
-        assert!((result[0].lower - (-2.0)).abs() < 0.1);
-    }
-
-    #[test]
-    fn test_keltner_channels_next_error() {
-        let mut kc = KeltnerChannels::new(14, 10, 2.0).unwrap();
-        let candle = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-
-        // The next method should return an error as it's not implemented
-
-        // Test that the next method works with a candle
-        let result = kc.next(candle);
-
-        // Since we haven't accumulated enough candles yet, we should get None
-        assert!(result.is_ok());
-        if let Ok(value) = result {
-            assert!(value.is_none());
-        }
-    }
-
-    #[test]
-    fn test_keltner_channels_reset() {
-        let mut kc = KeltnerChannels::new(3, 3, 2.0).unwrap();
-
-        // Add some candles
-        let candle1 = Candle {
-            timestamp: 1,
-            open: 10.0,
-            high: 12.0,
-            low: 9.0,
-            close: 11.0,
-            volume: 1000.0,
-        };
-        kc.next(candle1).unwrap();
-
-        // Reset
-        kc.reset();
-
-        // After reset, we should have cleared state
-        assert!(kc.current_ema.is_none());
-        assert!(kc.current_atr.is_none());
-    }
-}
