+++ src/indicators/volume/mod.rs
@@ -1,0 +1,16 @@
+//! Volume-based indicators
+//!
+//! This module contains volume-based indicators like OBV, Volume Rate of Change, and A/D Line.
+//!
+
+// Module declarations
+pub mod adl;
+pub mod cmf;
+pub mod obv;
+pub mod vroc;
+
+// Re-exports
+pub use self::adl::Adl;
+pub use self::cmf::Cmf;
+pub use self::obv::Obv;
+pub use self::vroc::Vroc;
